Various fixes

+++ src/main/java/slimeknights/tconstruct/library/modifiers/modules/armor/ArmorWalkRadiusModule.java
@@ -59,11 +59,11 @@ public interface ArmorWalkRadiusModule<T> extends ArmorWalkModifierHook, Modifie
   default void onWalk(IToolStackView tool, ModifierEntry modifier, LivingEntity living, BlockPos prevPos, BlockPos newPos) {
     if (living.onGround() && !tool.isBroken() && !living.level().isClientSide) {
       T context = getContext(tool, modifier, living, prevPos, newPos);
-      float radius = Math.min(16, getRadius(tool, modifier));
+      int radius = Math.min(16, (int) getRadius(tool, modifier));
       MutableBlockPos mutable = new MutableBlockPos();
       Level world = living.level();
       Vec3 posVec = living.position();
-      BlockPos center = new BlockPos(posVec.x, posVec.y + 0.5, posVec.z);
+      BlockPos center = new BlockPos((int) posVec.x, (int) (posVec.y + 0.5), (int) posVec.z);
       for (BlockPos pos : BlockPos.betweenClosed(center.offset(-radius, 0, -radius), center.offset(radius, 0, radius))) {
         if (pos.closerToCenterThan(living.position(), radius)) {
           walkOn(tool, modifier, living, world, pos, mutable, context);
